requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
requests
groq
ddgs
beautifulsoup4
pdfminer.six
sentence-transformers
faiss-cpu
numpy